Rename directory "test" to "blackbox" in monkey_island/cc/resources/
