+++ monkey/monkey_island/cc/resources/blackbox/__init__.py
