+++ monkey/monkey_island/cc/resources/blackbox/clear_caches.py
