+++ monkey/monkey_island/cc/resources/blackbox/log_test_endpoint.py
