+++ monkey/monkey_island/cc/resources/blackbox/monkey_test_endpoint.py
