+++ monkey/monkey_island/cc/resources/blackbox/telemetry_test_endpoint.py
