+++ monkey/monkey_island/cc/resources/blackbox/utils/telem_store.py
